[MEDIUM] Scheduler: add routine for regular inspection of the spool
This goroutine wakes up every minute and inspect the content of the spool
directories. For now, it only looks for actions in the new action directory, and loads
or removes actions based on scheduled and expiration dates. In the future, this
routine will also check the inflight and done directories to cancel timedout actions
and commands, and do more cleanup.