[major] scheduler refactoring for performance improvements
This rewrite adds supports for batch operations in a number of places, particularly
in the database where commands are now inserted in batches, and actions
are updated a lot less often when many commands return at once.
It also removes two commands directories, and uses Go channels to pass the
commands instead. This alone reduces disk IOs and processing time to marshal
and unmarshal JSON.
There should be a net gain in reliability due to the reduced pressure on the file
system, at the cost of a light increase in memory usage.