[major] PGP authentication of API clients, includes client code refactoring
This is a major change to the client code and the API. All API calls now require
a HTTP header called X-PGPAUTHORIZATION that contains a signature from a
known investigator. This work is also the opportunity to split up the API client
part of the MIG console into a new package called mig/client. This will facilitate
code reuse, particularly for standalone command line utilities.