[medium] agents runs modules in goroutines, not exec()
This patch changes the way modules are executed by agents.
Instead of running them as separate processes, invoked by
passing module parameters over the command line, they are
now launched as goroutines of the main agent process.
As a side effect, command line args length are not a problem
anymore.