[medium] Prevent command results spoofing by rogue agents
This issue was surfaced in https://bugzilla.mozilla.org/show_bug.cgi?id=1136816
The fix verifies that the command results contain the correct agent
queueloc value, which includes the agent secret. A rogue agent does not
know the queueloc value and thus will send spoofed results that do not
verify the condition of the new WHERE clause.