[major] support stdin parameter submission to modules in agent
This change is to support cases where parameters for modules could
exceed the limit for command line arguments. Modules can register
to recieve parameters over stdin.

Right now the code for reading the stdin based parameters is
embedded in the module. This could probably be centralized into
a utility function, but was initially left this way since all
modules right now are responsible for parsing their own input.

One potential problem case could occur where a module does not
read the input and the agent ends up blocking on the pipe. It may
be useful to implement some form of timeout on the writes to ensure
this does not end up causing the agent to stop responding.