[minor] better description on gpg key not found on client
When the .migrc is missing, a new one can be created, and if gpg is
installed, but there is no suitable key in the secring file, it will
display an error saying "No suitable key found". This commit also adds
to the error description the file which was inspected.

This fixes issue #132.