[medium/bug] terminate scheduler when heartbeat to relays fails, fixes #146
If RabbitMQ suffers a network partition, the scheduler will remain in a
half broken state where no message can be consumed or published, but
other goroutines work fine. This patch introduce a new dummy event that
heartbeats the scheduler and terminates it if the publication of the
heartbeat fails. The hearbeat is a normal event that expires in rabbitmq
after 100 minutes, like any other event. It is published to a queue that
could eventually be consumed for monitoring.