[minor] remove agt serviceInstall flag and add agent check in loader
Remove the service install flag to the agent as it's not really
neccessary; the standard daemon invocation will achieve the same result.

Also add a check in loader runs to see if the agent is running even if
no changes were needed. If not, try to start the agent. This handles
cases where the loader doesnt need to make changes, but the agent is not
running for some reason, and prior the loader would not attempt to
recover this scenario.