[minor] if the AMQP channels closes in getCommands, terminate
If the AMQP channel used in getCommands is closed due to an underlying
error in the AMQP connection, the goroutine exits. The heartbeat process
will also encounter the connection error during publish, and this will
result in termination of the agent.

Depending on the configured heartbeat time, this can take a while to
occur. Rather than wait for the heartbeat publish to fail, if we are
unable to recieve messages from AMQP also send a termination note here.