[minor/bug] use sync.WaitGroup in computeAgentsStats()
previous implementation had potential to deadlock function, as it was
possible to enter the wait loop before routines had been incremented,
since routines was being incremented inside the goroutine. this would
result in the function never exiting the loop comparing completed to
routines, as on the first iteration routines would be 0 and completed
would be 1.