[minor] fixes to agent/loader packaging, primarily for lintian errors
This fixes lintian errors reported for deb packages, which cause the
package installation to fail in some cases where frontends to dpkg are
being used to install.

Changes primarily applied to loader/agent packages. deb-server package
was not modified as the target installs to /opt, which is reported as an
error by the linter. Since a change to the install path is required to
fix this just leave as is for that target for now.

Closes #197