[minor] add support for user defined target macros
This provides the ability for users to define target macros in ~/.migrc,
as an example rather than requiring -t "status='online'", given a
section such as:

[targets]
    macro = allonline:status='online'

in .migrc, you can then do -t allonline. The intended specification
format is macro = name:<target string>. Multiple "macro" items can be
specified in the targets section.

This also removes the default target in mig-cmd, and the -t flag is
now a required option. This better aligns with how mig-console works.