[minor] agent epm implementation
EPM (extra privacy mode) is an agent option that if enabled, will
inform modules they should apply additional masking on returned results.

For example, by default if the file module matches a search, the name of
the matching file will be returned. With this enabled, it may inform the
investigator that something on the system matched, but will not provide
meta-data.