[medium] enhance support for multiple action signatures
Adds functionality to mig-cmd in addition to various bug fixes to
correctly support multiple signature verification with ACL module
weights.

This implementation is rudimentary in the sense it gives an investigator
the ability to output a signed action, which can be handed off to
another investigation for application of a second signature and
execution in the system.