Add persistent module audit
audit is a persistent module that uses libaudit-go
to setup audit rules(json format) and retrieves the
audit messages. It further correlates them to pack
in a single audit event similar to github.com/gdestuynder/audisp-json
Then it forwards them to an output medium.
Currently a http server that accepts POST is
implemented as an output medium.