[minor] in findOSInfo(), dont log normal errors at Info
On some distributions lsb_release will not be present. Currently when
the agent is run on these systems, an error is always generated on
startup after lsb_release cannot be found and the agent uses issue for
distribution detection.

This changes the function so it does not generate these errors as they
are expected during normal execution, but will still print them at the
debug level.

Closes #326