[minor/bug] register endpoints on status socket once
If the agent fails to bind the status socket, it will wait and retry
again. As part of this retry, it was also binding endpoints to register
them in the http package for the status socket which would result in a
panic.

Changes behavior so endpoints are only registered once.