[minor] findPublicIP: reduce Dial timeout to 5 seconds
Reduce the dial timeout to 5 seconds time 10. This operation should
normally occur quickly, but in some cases (e.g., an unreachable proxy in
use) the request will stall. This ensures we wait a maximum of 5 seconds
before trying another option.

This also aligns the timeout with the dial timeout for AMQP connections.