abstract Client http.Client with an interface
Introduces a new Client interface type which is used to interact with
the MIG API. The main intent of this interface is to allow substitution
of the Client type with other implementations (e.g., a Client type
suitable for unit testing).

The original Client becomes MIGClient, and satisfies the Client
interface.

A test file has been included that adds a new TestClient type which also
satisfies Client, and can be extended for internal tests.

Also, the existing http.Client member of the MIG client type has been
replaced with a requestor interface. This is intended to permit the
default client http.Client and transport configuration to be replaced if
custom options are desired.