mig-cmd: handle SIGINT nicer when interrupting an action
Use WaitGroups to avoid interleaving text if an action is cancelled.
This did not occur all the time, but this change should provide for more
consistent behavior.

mig-console didn't handle this at all; add a comment around this as well
and adjust mig-console to provide the additional extra parameter
required.