add docker-compose based demo environment
The existing MIG docker image can be used as a standalone demo; this
change extends the image to be used in a compose configuration, creating
separate containers for each component.

The compose environment is intended to potentially be used for
deployments other than a demo, by selecting individual components
intended for execution and configuring these components using the
environment.