simplify ACL type and remove Permission, fix authorization bypass
This change removes the Permission type, and modifies ACL to simply be a
map of a given module name and the authorization specification.

It was previously stored as a list of ACL types, with each type being a
map which was not required.

This also fixes a bug in authorization, where an action would be
authorized if the first operation validated against the ACL. This means,
it was possible if an action contained more than one operation, an
investigator could be authorized for execution if the first module
matched the ACL. Note that for this to happen, the first operation
checked would need to be authorized and in general deployments this does
not have a significant impact.

Resolves #394