Pass trimmed input line to commands
The space was trimmed and split in console.go, but the original `input`
was passed to the specified command.

The command then splits the input again, but doesn't necessarily trim
trailing spaces that are then counted as empty arguments. This breaks
the check for required argument when using tab completion.